+++ src/main/java/com/supercoding/hanyipman/service/OptionService.java
@@ -11,10 +11,9 @@ import com.supercoding.hanyipman.error.domain.MenuErrorCode;
 import com.supercoding.hanyipman.error.domain.OptionErrorCode;
 import com.supercoding.hanyipman.error.domain.OptionItemErrorCode;
 import com.supercoding.hanyipman.error.domain.ShopErrorCode;
-import com.supercoding.hanyipman.repository.MenuRepository;
+import com.supercoding.hanyipman.repository.menu.MenuRepository;
 import com.supercoding.hanyipman.repository.OptionItemRepository;
 import com.supercoding.hanyipman.repository.OptionRepository;
-import com.supercoding.hanyipman.repository.shop.ShopCustomRepository;
 import com.supercoding.hanyipman.repository.shop.ShopCustomRepositoryImpl;
 import com.supercoding.hanyipman.security.JwtToken;
 import lombok.RequiredArgsConstructor;
